+++ commands/command_base.h
