+++ commands/command_edit_trip.cpp
