+++ commands/command_edit_trip.h
