+++ commands/command_private.cpp
