+++ commands/command_private.h
