+++ desktop-widgets/subsurfacewebservices.cpp
@@ -4,8 +4,8 @@
 #include "core/webservice.h"
 #include "core/settings/qPrefCloudStorage.h"
 #include "desktop-widgets/mainwindow.h"
-#include "desktop-widgets/command.h"
 #include "desktop-widgets/usersurvey.h"
+#include "commands/command.h"
 #include "core/trip.h"
 #include "core/errorhelper.h"
 #include "core/file.h"
