+++ profile-widget/profilewidget2.cpp
@@ -24,8 +24,8 @@
 #include "desktop-widgets/diveplanner.h"
 #include "desktop-widgets/simplewidgets.h"
 #include "desktop-widgets/divepicturewidget.h"
-#include "desktop-widgets/command.h"
 #include "desktop-widgets/mainwindow.h"
+#include "commands/command.h"
 #include "core/qthelper.h"
 #include "core/gettextfromc.h"
 #include "core/imagedownloader.h"
